==> data_processing/3_polling_station_village.ipynb <==
import pandas as pd
import geopandas as gpd
from shapely.geometry import Point
import warnings


polling_stations = pd.read_csv('data/polling_stations_with_coordinates.csv')
village_areas = gpd.read_parquet('maps/hualien_division_area_village.geoparquet')

village_areas["number"] = village_areas.primary_name.str[0:2].apply(lambda x: polling_stations[polling_stations.village_neighborhood.str.contains(x)].number.to_list())

new_village = village_areas.explode("number")

merged_df = (
    polling_stations
    .merge(new_village, on="number", how="left")
    .rename(columns={
        "primary_name": "area_name",
        "number": "polling_station_id",
        }
    )[['polling_station_id', 'address', 'longitude', 'latitude', 'area_name']]
)

na_filler = merged_df.query("area_name.isna() & address.str.contains('北昌')")
merged_df.loc[na_filler.index, "area_name"] = "北昌村"
merged_df.to_csv("./data/polling_station_area.csv", index=False)


# merged_df.query("area_name.isna()")

polling_stations.query("region == '花蓮市'")

# voting_data = pd.read_csv('./data/recall_vote_data_detailed.csv')
# voting_data.query("選舉區 == '花蓮市'")
