==> data_processing/5_read_legislator_polling_results.ipynb <==
import pandas as pd



legislator_polling_results = pd.read_excel(
    "./data/legislator_election_results.xlsx",
    skiprows=6,
    header=None,
    names=[
        "region",
        "village",
        "polling_station_id",
        "許願神",
        "張美慧",
        "傅崐萁",
        "周玉梅",
        "effective_vote",
        "invalid_vote",
        "total_vote",
        "pickup_no_vote_count",
        "pickup_vote_count",
        "remaing_vote_count",
        "total_voter_count",
        "voting_rat_perct"       
    ]
)

legislator_polling_results['village'] = legislator_polling_results.village.fillna("總共")
legislator_polling_results["polling_station_id"] = legislator_polling_results.polling_station_id.fillna(0)
legislator_polling_results.region = legislator_polling_results.region.fillna(method="ffill")

legislator_polling_results.to_csv("./data/legislator_polling_results_cleaned.csv", index=False)
